crates/fitswcs-sys: import bindings to cfitsio/wcslib
These will support the image cutout Lambda service. We vendor in
the source to cfitsio and wcslib because in this very targeted use
case, there's no point in trying to be generic about things. In
retrospect we could have possibly relied on existing generic `-sys`
crates; I wanted wcslib to be able to link to cfitsio if needed, but
I think it doesn't actually need to do so.