src/s3buffer.rs: fix an error with undersizing the buffer
If we made a request that was for more bytes than could fit in the
buffer, we'd stop at the buffer size and return an error. Oops! If we make sure
to request the full number of bytes, the buffer will grow and it will be OK.

I think I'm only seeing this because of a file where the three-tier
caching scheme isn't quite working, and a request for image data is
landing on buffer "B" instead of buffer "C" as intended. This isn't
ideal, but seems fine, so I don't see a need to try to fix that. If we
need to grow the buffer, that's definitely what we should do.