Figured out the proxy response magic?
I think I may have figured out the right magic to obtain control over
the HTTP response headers, status code, etc., in this framework. It
requires sacrificing the "bare" version of the program, but I have no
need for it anymore, anyway.

With this in hand I think that we could in fact deliver non-JSON outputs
from our API endpoints, but I don't want to monkey with anything
existing at this late hour.