Provide JSON schemas for the response types
These are needed for the automagical AWS API documentation framework.

We also slightly alter the presign API to return the redirect location
as JSON body content, while continuing to return a redirect code with
the appropriate `location` HTTP header. This allows the "oneshot" form
of the API invocation to be more useful.