json-schemas/mosaic_package_response.json: workaround for orval bug
In the Starglass frontend, the `orval` API processor has some kind of
bug where it produces invalid syntax when given an input schema that
specifies a dictionary/object that can contain anything. In my testing,
the only way we can avoid this is to explicitly specify at least one
potential property of the object. Fine.