Return useful HTTP error codes
It turns out that if anything ever went wrong during our processing, we
would always return an HTTP 500 error in the lambda, which almost surely
crashed the lamba runner too. But for cases where, say, an input declination
is 100 degrees, we should obviously be returning a 400 Bad Request error
and not actually crashing.

Now we start doing that. This is done with a helper `HttpExposedError`
class, which the high-level dispatcher notices and translates into an
appropriate HTTP error response. A fun little extension trait for
Option<> provides some nice ergonomics for common cases.